+++ CHANGELOG.md
@@ -9,6 +9,7 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 - Immediate retries to RetryInterceptor, defaults to 3 before applying wait period between retries
 - Support for Http Methods DELETE, PATCH, POST and PUT.  REST Adaptor and REST Client generators now 
 support RAML that contains asynchronous DELETE, PATCH and PUT, and synchronous PATCH, POST and PUT.
+- Support for matchers in TypeCheck class.
 
 ### Changed
 - Optimistic lock log message changed from warn to debug.